CMYK Support for SetColor functions
Suppor for use CMYK colors in SetDrawColor(), SetFillColor() and SetTextColor().
If you pass no parameter, you get black.
If you pass 1 parameter, it is a grayscale component (between 0 and 100).
If you pass 3 parameters, they are RGB components (between 0 and 255).
If you pass 4 parameters, they are CMYK components (between 0 and 100).
Can you pass the color values as an array:
$rgb_color = [$red, $green, $blue];
$cmyk_color = [$cyan, $magenta, $yellow, $key]; // In CMYK the black color is called Key.